Struct google_kgsearch1::EntitySearchCall

pub struct EntitySearchCall<'a, C, A> where C: 'a, A: 'a {
    // some fields omitted
}

Searches Knowledge Graph for entities that match the constraints. A list of matched entities will be returned in response, which will be in JSON-LD format and compatible with http://schema.org

A builder for the search method supported by a entity resource. It is not used directly, but through a EntityMethods instance.

Example

Instantiate a resource method builder

 
// You can configure optional parameters by calling the respective setters at will, and
// execute the final call using `doit()`.
// Values shown here are possibly random and not representative !
let result = hub.entities().search()
             .add_types("nonumy")
             .query("dolores")
             .prefix(false)
             .limit(-95)
             .add_languages("aliquyam")
             .indent(false)
             .add_ids("no")
             .doit();

Methods

impl<'a, C, A> EntitySearchCall<'a, C, A> where C: BorrowMut<Client>, A: GetToken

fn doit(self) -> Result<(Response, SearchResponse)>

Perform the operation you have build so far.

fn add_types(self, new_value: &str) -> EntitySearchCall<'a, C, A>

Restricts returned entities with these types, e.g. Person (as defined in http://schema.org/Person).

Append the given value to the types query property. Each appended value will retain its original ordering and be '/'-separated in the URL's parameters.

fn query(self, new_value: &str) -> EntitySearchCall<'a, C, A>

The literal query string for search.

Sets the query query property to the given value.

fn prefix(self, new_value: bool) -> EntitySearchCall<'a, C, A>

Enables prefix match against names and aliases of entities

Sets the prefix query property to the given value.

fn limit(self, new_value: i32) -> EntitySearchCall<'a, C, A>

Limits the number of entities to be returned.

Sets the limit query property to the given value.

fn add_languages(self, new_value: &str) -> EntitySearchCall<'a, C, A>

The list of language codes (defined in ISO 693) to run the query with, e.g. 'en'.

Append the given value to the languages query property. Each appended value will retain its original ordering and be '/'-separated in the URL's parameters.

fn indent(self, new_value: bool) -> EntitySearchCall<'a, C, A>

Enables indenting of json results.

Sets the indent query property to the given value.

fn add_ids(self, new_value: &str) -> EntitySearchCall<'a, C, A>

The list of entity id to be used for search instead of query string.

Append the given value to the ids query property. Each appended value will retain its original ordering and be '/'-separated in the URL's parameters.

fn delegate(self, new_value: &'a mut Delegate) -> EntitySearchCall<'a, C, A>

The delegate implementation is consulted whenever there is an intermediate result, or if something goes wrong while executing the actual API request.

It should be used to handle progress information, and to implement a certain level of resilience.

Sets the delegate property to the given value.

fn param<T>(self, name: T, value: T) -> EntitySearchCall<'a, C, A> where T: AsRef<str>

Set any additional parameter of the query string used in the request. It should be used to set parameters which are not yet available through their own setters.

Please note that this method must not be used to set any of the known paramters which have their own setter method. If done anyway, the request will fail.

Additional Parameters

bearer_token (query-string) - OAuth bearer token.

pp (query-boolean) - Pretty-print response.

prettyPrint (query-boolean) - Returns response with indentations and line breaks.

upload_protocol (query-string) - Upload protocol for media (e.g. "raw", "multipart").

access_token (query-string) - OAuth access token.

fields (query-string) - Selector specifying which fields to include in a partial response.

quotaUser (query-string) - Available to use for quota purposes for server-side applications. Can be any arbitrary string assigned to a user, but should not exceed 40 characters.

callback (query-string) - JSONP

oauth_token (query-string) - OAuth 2.0 token for the current user.

key (query-string) - API key. Your API key identifies your project and provides you with API access, quota, and reports. Required unless you provide an OAuth 2.0 token.

uploadType (query-string) - Legacy upload protocol for media (e.g. "media", "multipart").

alt (query-string) - Data format for response.

$.xgafv (query-string) - V1 error format.